8569 Ability to easily extract email attachments
Code + unit for feature request 8569-Ability to easily extract email attachments

Added an option to SplitAttachmentsExpression to have the attachments be split into
new messages with a byte[] body that contains the attachment, instead of a single
attachment on a clone of the message.

NOTE: Default behaviour is retained (i.e. splitting into cloned messages each with a
single attachment).

Former-commit-id: b97c2ba5bae3c5a17f1f006247c6cbec5c01b48e