CAMEL-9572: Validator: clearSchemaCache method refactored
- it is not a good idea to re-read the schema in the clearSchemaCache
method. Therefore the readSchema method is moved to the SchemaReader.
Now, if the schema was updated and the clearSchemaCache method is called
then the re-reading is tried during the first process call. This change
was especially necessary  to support the case where the schema document
was deleted and the clearSchemaCache was called. In this case. now no
error occurs during the clearSchemaCache call, but the error "no file
found" occurs during the next process call. This is the expected
behavior, in the previous solution the cache would not be cleared at all
and the processing would have went on. 

Change-Id: Ie206872875e2bddaeb036fc4dfb43ec907770787

Former-commit-id: 77e079a83861b93d39d556a76019e78cfb9389db