camel-xmpp: fix private chat response issue
Responding to a private chat wasn't able because the message was always sent to

    String participant = exchange.getIn().getHeader(XmppConstants.TO, String.class);

However, the `XmppConstants.TO` in such messages is the camel instance
(which received the message) and not the chat's other participant.
Thus, responding to a chat message always resultet in a message to itself.
This patch fixes that issue and addresses messages always to

    String participant = endpoint.getParticipant();


Former-commit-id: f4146a7594bdc6d22ae392afc203f3a876c190ed