Fix over writing previous addToResult calls
I noticed that multiple calls to the addToResult method in the org.apache.camel.component.aws.ddb.AbstractDdbCommand within the child class org.apache.camel.component.aws.ddb.ScanCommand caused the next call to over write the previous calls to addToResult in the header results.  The end result is only having the last result value returned in the out message which is the SCANNED_COUNT header.

The exchange.getOut() already passes back the "In" Message or creates a new Message object if no out message exists does not need to copy over the "In" message into the "Out" message.

Former-commit-id: 45c375c38e1a02347c17d172a6bd7dc68aebcdd4