Add publishedEndpointUrl capability to rsServer
Endpoint definition allows to specify publishedEndpointUrl, but this attribute is not available in rsServer element. This attribute (publishedEndpointUrl) is used by WadlGenerator to compose baseUri resource in the WADL. If you are deploying a REST service in an internal net but exposed to internet (through reverse proxy, for example) , WADL should inform such publishing address (in a similar way that you can do it with WSDL in camel-cxf).

Actually, since no publishedEndpointUrl can be used in rsServer definition, baseUri is built with HTTP request information. That means internal network information could be leaked in a reverse proxy scenario.

Former-commit-id: 46699f94c0f0bb4c2300b8650e78bc6ffc03377a