Update Camel Salesforce integration tests
Tests in SalesforceComponentConfigurationTest were not updated to the
latest functionality provided by the
SalesforceComponent::completeEndpointPath.

Also renamed SalesforceComponentConfigurationTest to
SalesforceComponentConfigurationIntegrationTest, the test requires
Salesforce access to run. Renamed the `salesforce-test` profile to
`integration` and enforced existence of
`test-salesforce-login.properties` file in the root of
`camel-salesforce-parent` project -- it is needed for the integration
tests and contains the credentials required for accessing Salesforce
account used.

Details in README.md.


Former-commit-id: 52459d26ae17bb0436fb84354b169e8a649e4f3e