CAMEL-10679
Producer does not populate attachments from response.

I would think of refactoring the whole part of the response handling.
In my opinion it would be best to handle this the same way as the
BasicMessageFilter does this for the request message.
This way anyone could implement custom behaviour if neccessary.
I would have done it, but i didnt want to break anybodys code, so i just
implemented it the same way the headers are populated.
Feel free to contact me about the change.


Former-commit-id: f20d4a606d2f3249d3fd77af58922103d37a947f