CAMEL-9945 Upgrade to jetty 9.3
Java compiler uses the actual return type of the method in the compiled
class and as the signature of
`org.eclipse.jetty.client.HttpClient::getProtocolHandlers` method
changed between Jetty 9.2 and 9.3 in return type it could be compiled
and run against both versions, but it could not be compiled with one
version and run against another.

This commit uses reflection when calling `getProtocolHandlers` method to
maintain compatibility with 9.2 and 9.3 versions of Jetty regardless of
compile time vs runtime version.


Former-commit-id: 8392379c320aa6de48acf0a07105c70148487ff8