CAMEL-10759 RabbitMQ Component binding args
Introduced 3 new Endpoint URI properties 
- exchangeArgs
- queueArgs
- bindingArgs

As an easier way to specifiy rabbitmq properties at declare time.
Meanwhile deprecated the ArgsConfigurer class whilst maintaining backwards compatibility. It's actually possible to use both styles of arg declaration because the Maps are combined. 

Introduced a new integration test for the bindingArgs which tests a headers exchange correctly routing to a queue. Updated the previous integration test RabbitMQInOutIntTest to use the new style of arg when setting the "x-expires" property on the queue.

Finally added a small example of how to declare a headers exchange using new feature. 

Former-commit-id: c22d808a21c256acdc0f05594d8d53de005a9138