CAMEL-10764 The use of BrowserCompatHostname...
...Verifier is deprecated

The BrowserCompatHostnameVerifier has been deprecated in HttpClient
version 4.4, instead DefaultHostnameVerifier should be used. The commit
also changes the type of `x509HostnameVerifier` from
`X509HostnameVerifier` to `HostnameVerifier`.

Due to HTTPCLIENT-1802[1] version of HttpClient was upgraded to 4.5.3
from 4.5.2.

Compatibility note: this changes the getter `getX509HostnameVerifier`
from `X509HostnameVerifier` to `HostnameVerifier` so any client code
using this getter needs to be changed. For the setter
`setX509HostnameVerifier` this is not an issue as the deprecated
`X509HostnameVerifier` extends from `HostnameVerifier`.

[1] https://issues.apache.org/jira/browse/HTTPCLIENT-1802


Former-commit-id: bcc99170cbe81031dbd4fa7b32ce36c0e5a60ba8