CAMEL-10849: Salesforce: subscription channel ...
...created per component

This changes the way configuration, specifically `initialReplayIdMap`
and `defaultReplayId` are configured on the CometD client.

The Bayeux CometD client is left at the Component level, this is to
limit the number of streaming clients connected to Salesforce as there
are limits on those.

The SubscriptionHelper was previously tied to the topic it subscribed to
on creation, removes that constraint. This allows per-endpoint
customization of the client (subscriptions) on subscription from
consumers.


Former-commit-id: f507f4eed27a54dbfb30f019e7bbd3f8b4483920