CAMEL-10966 Salesforce Maven Plugin doesn't esc...
...ape strings when doing the camel-salesforce:generate phase

This changes the test a bit, let me explain how: there is no need to use
reflection to make a method in the object under test accessible, we can
simply change the visibility of the target method and refactor where
needed. This makes the object under test easier to test and in the end
better in quality.

So I changed the visibility of the `engine` field to package private to
make it accessible from the test, and refactored the initialization of
`VelocityEngine` into `createVelocityEngine` method so that it can be
used in the test. Now this can be used in the `setUp` method of the
test.

To clean up the test a bit I used `Parameterized` runner which is nice
for tests that repeat the same checks over a number of inputs --
parameters, there is a unused field `json` but that's the way the
`Parameterized` runner works: every parameter needs to be used. I've
added `@Parameters(name = "json = {0}, source = {2}")` to make the tests
cases for different parameters more distinct from each other, with IDE
support it is possible to run all or only single test case (set of
parameters)[1].

I also added a JUnit rule that will create and cleanup the temporary
directory, that way the test can be run via different runners in
parallel -- for instance I use Infinitest that runs tests in the IDE in
the background, now I'm sure that those won't interfere with the tests I
run manually. So this makes the test run less prone to outside
influences and less _flaky_.

[1] http://blog.moritz.eysholdt.de/2014/11/new-eclipse-junit-feature-run-subtrees.html


Former-commit-id: cccf6726b1fb59d592453e32909c0c3fa39dad2c