CAMEL-11025 RestProducer should set Accept and ...
...Content-Type headers if given in endpoint configuration

This sets the headers on the incoming message, if not already set, to
the supplied values from the endpoint configuration. Endpoint property
`consumes` is mapped to HTTP `Accept` header, and endpoint property
`produces` is mapped to HTTP `Content-Type` header; both are passed as
Camel message headers to the delegate producer that will carryout the
request.


Former-commit-id: ab6fb50abdf0535204e395fe97007a2d34e0cbfa