CAMEL-10938 Camel Salesforce : add an option to...
...retrieve login information for testing purpose from env var or
system properties

This introduces `authenticationType` Salesforce component property that
can be used to force, as opposed auto-determine authentication flow the
component should perform.

It ties into the integration tests setup as it enables tests to be run
with any supported authentication type.

More to the point of the initial JIRA, this change enables configuring
the integration tests using environment variables, Java system
properties or the established `test-salesforce-login.properties` file.

The syntax in the `test-salesforce-login.properties` has changed
slightly to better accomodate the environment variable syntax: camelcase
was dropped in favor of dot notation.


Former-commit-id: c63a9fd235120c2b5b845e0583428b9c149a0769