CAMEL-11194 Upgrade Checkstyle and update rules
This updates the Checkstyle version to 7.6.1, it also changes the
execution id from `validate` to `default-cli` which allows the
Checkstyle Maven plugin to be run with just `mvn checkstyle:check`.

The change in Checkstyle rules now allows for different header comment
on source files and on properties files. The rule that affects the use
of `final` keyword and the rule that checks the `return` statement count
per method were tweaked to reflect the current code base.

A `license-maven-plugin` from `com.mycila` was added within a
`update-license` Maven profile that can be used to update the license
headers in files with simply:

    mvn -Pupdate-license


Former-commit-id: c96b417942338355c162701712de3f61534ce52d