CAMEL-11261 Revise Camel context destruction in...
... Spring (Boot) applications

This changes the way `CamelContext` is started/stopped when used with
Spring Framework. It uses `ContextRefreshedEvent` for startup and
`Lifecycle::stop` to stop, the reason why `Lifecycle` can't be used for
startup is that only beans that implement `SmartLifecycle` are started
and `SpringCamelContext` can't implement `SmartLifecycle` because of
overlapping `isAutoStartupMethod` in `RuntimeConfiguration` and
`SmartLifecycle` with different return types -- that cannot be used as
covariant return type as the `SmartLifecycle` signature uses primitive
type.

The order of startup is quite important, as some changes to Camel
context are not properly expressed if made after context startup (for
instance `EndpointStrategy` change), so `SpringCamelContext` has
`LOWEST_PRECEDENCE` and `RoutesCollector`s, `CamelContextFactoryBean`
have +1 precedence over that in order to add routes and setup context
before startup.

Using `ContextRefreshedEvent` is actually better than using
`SmartLifecycle` as `SmartLifecycle` invocation precedes
`ContextRefreshedEvent` emission, so there could be beans in the context
that also start on `ContextRefreshedEvent` and Camel would start before
those.

One side note on using events is that at some point Spring might
default, however unlikely that seems, to using multi-threaded event
listener invocation, so this might complicate things.


Former-commit-id: fbf0ffad627b42088705afe0e29c5ebc1a77809a