CAMEL-11667: Support Metadata component extension
Adds support for `MetaDataExtension` in the Salesforce component.
Currently two kinds of metadata are supported:
 - listing Salesforce Object types
 - listing fields of Salesforce object type

For Salesforce object types object name and label are returned and for
Salesforce object fields name and field type are returned.

New facade to use low level Salesforce RestClient was added
`SalesforceClientTemplate` intended to be used from extensions such as
this and the existing `ComponentVerifierExtension` was refactored to use
it.


Former-commit-id: ce86029bb906cb0db78c3b22f84c01aa2da2e587