CAMEL-11814: makes NO_START a ThreadLocal, and ...
...changes the way CamelMainRunController is started

We need to change NO_START flag back to ThreadLocal as there is a use
case when it's used from a single classloader outside of the tests in
wildfly-camel.

CamelMainRunController would atempt to start CamelContext from a thread
that does not have the NO_START flag defined (it's a ThreadLocal), so
it can only run when CamelContext is started. It's main purpose is to
prevent the SpringBoot application JVM from terminating so having it
run when the CamelContext is started doesn't prevent that.


Former-commit-id: 1c17aa3f3c4cde4862bee11698c294f543182ae0