Update xquery-component.adoc
It took me quite a while to figure out that custom functions in Camel 2.18, 2.19 and 2.20 could result in a NullPointerException, so I think it could save others some time and frustration if there is a note like the one i propose in the documentation. I can not for sure promise that this will work in 2.21, but based on the discussions in PR CAMEL-12081, it looks like it will be fixed there.

Former-commit-id: 5ed12728ebcbddc7de1ffdfb1a1dda8ede029c1b