Make olingo2 test setup deterministic
`camel-olingo2-component` needs a generated sample project for tests,
this is done in a separate profile that is active by default, this poses
an issue if the generated project already exists. This adds
maven-clean-plugin execution to remove that sample project if it already
exists.
Now the test should be deterministic and should not fail the build
regardless of the existence of the sample project.


Former-commit-id: 95f5438fbc447f529ba235d1239b0edecc2bfc74