CAMEL-11930: camel-box and camel-linkedin make ...
...`test` goal execute successfully

`camel-box` and `camel-linkedin` components uses
`camel-api-component-maven-plugin` to generate sources, this fails if
javadoc is not present.
If run via `mvn install` this works as the javadoc dependency is present
in the reactor. If run with just `mvn test` this fails as it is not.
This moves the `maven-javadoc-plugin` to phase `generate-test-resources`
so that it will be present when running `mvn test`.


Former-commit-id: 20700b84f84f6fd1c37c929930a72f8120a5e800