CAMEL-11930: enable configuration override via ...
...environment variables

It could be easier to set environment variables in the job configuration
to configure the pipeline job. With this Jenkins node name/label
selector can be specified in `AGENT_LABEL`, likewise `JDK_NAME` is the
name of JDK installation in Jenkins and additionally `LOCAL_REPOSITORY`
environment variable can point to a local repository that the artifacts
are installed to.


Former-commit-id: 0abe9ee8a2b2a4f7463cd404a823bbbfab8c24db