Fix DropboxAPIFacade.java
Specifically the bug was in class DropboxAPIFacade.java lines 114, 115, 142, 143. The exception thrown is about upload a file into a not file path or a folder into a not folder path, but the new version of dropbox java sdk don't provides the method entry.isFile() or entry.isFolder(), and the code as is now, is throwing erroneous exceptions, if the file exists remotely, a exception will be thrown, and in config, we have the force mode, i think this snipped of code should be erased or the mode=force param needs to be verified.

Former-commit-id: df63f141ab829146a5586f2b8da43eb8119b536b