Make sure to check Maven result in daily build
In the daily build to run tests on JDK 9/10/11 we don't check for the
result of the Maven invocation. This might lead to tests not being run
due to a Maven dependency issue and the build outcome would be success.

This gathers result codes from invoking Maven and fails the build if
they do not end up being 0 for all 9/10/11 tests.


Former-commit-id: f2fc96756cc17126aa2c506234c2b1823614f9ea