CAMEL-12682: add support for auto discovery of ...
...Registry beans

This adds support for Spring beans of type `Registry` to be present in
the `ApplicationContext` and replace the default
`ApplicationContextRegistry`.

Multiple `Registry`-ies can be present in the `ApplicationContext` and
order preference can be expressed with standard Spring `Ordered`
contract.

This allows having a custom `Registry` and `ApplicationContextRegistry`
in the same `ApplicationContext` if needed.


Former-commit-id: b62825f404520ff006e058141419bdb541e2e794