CAMEL-12728: much better TLS/SSL configuration ...
...in `camel-rest-swagger`

Setting up TLS/SSL settings in `camel-rest-swagger` was limited to the
delegate component being configured and present in the context. This
adds support for setting TLS/SSL configuration via Camel's standard
`SSLContextParameters` and `useGlobalSslContextParameters` flag used in
many components that deal with TLS/SSL.

In addition to that, this adds a new parameter considered in the
`RestProducerFactory` SPI: now parameters that need to be applied to the
delegate component can be specified in the `component` key of the
parameter Map. Those properties will be only considered if the component
is not already present in the Camel context. That is, if the
`CamelContext::getEndpoint` would have the side effect of creating the
delegate component and adding it to the Camel context. Now that
component can be customized with properties from the `component`
parameters Map.


Former-commit-id: 93cda6ebc3416f0d24359da882ce7bb99dc087d4