CAMEL-13108: Move mock component out of camel-core (#2738)
Some work to get closer to be able to move mock, dataset and test components out of camel-core. However the builder stuff is tightly coupled to camel-core builder package and would require some more work to find a good way to do this. Will look at this later, but wanted to get this work so far upstream.


Former-commit-id: ff26402ba17373e06c37d230507170c434282761