Fix for CAMEL-13282
When a JAXB class that is created or generated has no @XmlRootElement
annotation JaxbDataFormat would try to find a ObjectFactory method which
can create a JAXBElement for the type of the object. It would then use
the JAXBElement to marshal the object.

While trying to find the objectFactory and to call the appropriate
method it would catch ALL Exceptions that are thrown and just logs a
message.

While i don't see why an Exception should only be logged here this
breaks when the exception is thrown due to a validation error. Because
the calls to the marshal operation will throw an MarshalException in
that case which would be caught by the catch block. In other cases where
no object factory is used but JAXB classes with XmlRootElement
Annotations or JAXBElements are used the MarshalExceptions will be
thrown and the doMarshal method will be left.

So i basically check the type of the exception that is thrown and
rethrow it if it's a MarshalException. All other exceptions are handled
in the same way as before. I kept the old exception handling in order to
keep the code as much backwards compatible as possible. But it should be
removed for 3.0.0 at the latest


Former-commit-id: 7c1e0a2e561eaa1d6d865863ea92441fbcc2e4d0