CAMEL-13386: Maven Flatten plugin refinements
This makes sure that Maven Flatten plugin is run in all profiles not
just for `release` and changes the `properties` element processing to
`keep` in contrast to the previous setting (`expand`) it should not
replace property placeholders with expanded properties that might
contain values specific to the machine the build is running.

Also makes sure that the distribution archives do not hold the
`.flattened-pom.xml` file.


Former-commit-id: c348cb6ca68265dcf7ca94ba3943d384e69d0e11