CAMEL-13496: set mavenOpts for Maven Invoker pl...
...ugin

As noted on the issue, setting `mavenOpts` only adds to the parameters
on the `java` command line, but luckily the last parameter present is
the one that is taken into account.

This sets the heap size to 64MB which should be enough for running
simple integration tests and prints the JVM configuration so if the
execution fails we can see the used configuration.

Also makes the `camel-restdsl-swagger-plugin` integration tests are only
run in the `integration-test` phase.


Former-commit-id: 1bd1e01ee3f3ee4fb63140149b1c66c0ffa3f223