CAMEL-13589: copy over examples
Antora requires that any `include::`-ed examples are in the `examples`
directory within a module[1]. To make this work for Camel we have to
copy the `.java` or `.xml` files referenced are copied over to the
module from the source tree of the Maven module.

Since several `.adoc` files both within the user manual and in the
component documentation reference files out of their own module using
relative paths to included files soon becomes an issue. That's why
having them absolute to the repository root makes more sense.

This adds changes to the `docs/gulpfile.js` to find all `include::`-ed
examples referenced from `.adoc` files and copy them over to the
corresponding module (user manual or component reference).

It also changes the paths in `include::`s to absolute for the above
stated reason.

Also makes the links point to the github repository via absolute URLs.

[1] https://docs.antora.org/antora/2.0/modules/#module-overview


Former-commit-id: bf2531bf17a853aa1ea499989d4c5baa72701fe8