CAMEL-13800: fix some issues in mobile resolutions
We have some text, notably section headings with long titles, that
cannot be broken into several rows. The default behavior in such cases
for many low resolution mobile browsers is to zoom out and while doing
that the text size is unreadable. This tries to reword some of the text
so it can be broken into multiple rows preventing this zoom out.


Former-commit-id: b6208d8fa9a017bf0edd3d1b3d88e0ef785a0f16