Camel-Salesforce: Support for relationship queries.
* Enhancements: Support for relationship queries – both child queries
and parent queries.

* Fix: Polymorphic fields can result in broken DTOs. E.g., if Contact and
Lead both have an External Id field, the POJO generator would generate
two fields with the same name. The solution is to have a `who` property
that accepts an AbstractDescribedSObjectBase instance.

* The "_Lookup" DTOs are no longer needed. You can now just use the
Class you’re looking up to, e.g. "Account" instead of
"Account_Lookup". This is a breaking change.

* Fixed: Tests unnecessarily referring to Custom Fields that are not
represented in the salesforce metadata.


Former-commit-id: 0cd6fa45fa7a13ee5ca8a13f837049dee4a82575