CAMEL-13981: logging in the annotation processor
This adds logging to the ModelAnnotationProcessor that ends up invoking
the CoreEipAnnotationProcessorHelper that should generate the
DefinitionPropertiesPlaceholderProviderHelper class.

The only way to see those messages is if their level is set to WARNING,
so this might trigger some tools to consider these as warnings.

With MCOMPILER-224[1] we should be also able to see the NOTICE output,
this is not the case in practice.

[1] https://issues.apache.org/jira/browse/MCOMPILER-224


Former-commit-id: 7c593199ebbec2ff4f17165c667e457362213ff9